Move all styles to an external stylesheet file
