+++ src/containers/FoodsList.js
@@ -4,41 +4,13 @@ import PropTypes from 'prop-types';
 import {
   Box,
   CircularProgress,
-  GridList, GridListTile, GridListTileBar, IconButton, ListSubheader, makeStyles, Typography,
+  GridList, GridListTile, GridListTileBar, IconButton, ListSubheader, Typography,
 } from '@material-ui/core';
 import { Link } from 'react-router-dom';
-
-const useStyles = makeStyles(() => ({
-  root: {
-    display: 'flex',
-    flexWrap: 'wrap',
-    justifyContent: 'space-around',
-    overflow: 'hidden',
-  },
-  circularProgress: {
-    padding: '50px',
-  },
-  gridList: {
-    width: '100%',
-  },
-  title: {
-    display: 'flex',
-  },
-  iconTop: {
-    background: '#d64646',
-    color: '#eeeeee',
-    marginRight: '10px',
-    marginBottom: '10px',
-  },
-  icon: {
-    color: 'rgba(255, 255, 255, 0.54)',
-    background: 'rgba(255, 255, 255, 0.14)',
-
-  },
-}));
+import styles from '../asset/styles';
 
 const FoodsList = (foodItems) => {
-  const classes = useStyles();
+  const classes = styles();
   return (
     <Box p={5}>
       <div className={classes.root}>
@@ -48,7 +20,6 @@ const FoodsList = (foodItems) => {
             <div className={classes.circularProgress}>
               <CircularProgress size={80} color="secondary" />
             </div>
-
           </Box>
             : <GridList cols={4} cellHeight={280} className={classes.gridList}>
         <GridListTile key="Subheader" cols={4} style={{ height: 'auto' }}>
